+++ bt/app/match_rule.h
@@ -4,9 +4,9 @@
 
 namespace bt {
     enum class match_scope {
-        any = 0,
-        domain = 1,
-        path = 2
+        any     = 0,
+        domain  = 1,
+        path    = 2
     };
 
     class match_rule {
@@ -18,6 +18,7 @@ namespace bt {
         std::string value;
         match_scope scope{match_scope::any};
         int priority{0};
+        bool is_regex{false};
         bool app_mode{false};
         std::string firefox_container;
 
@@ -30,5 +31,8 @@ namespace bt {
         static match_scope to_match_scope(const std::string& s);
 
         static bool parse_url(const std::string& url, std::string& proto, std::string& host, std::string& path);
+
+    private:
+        bool contains(const std::string& input, const std::string& value) const;
     };
 }